+++ app/src/main/java/com/vn/hcmute/team/cortana/mymoney/usecase/remote/PersonUseCase.java
@@ -11,6 +11,7 @@ import com.vn.hcmute.team.cortana.mymoney.exception.UserLoginException;
 import com.vn.hcmute.team.cortana.mymoney.model.Person;
 import com.vn.hcmute.team.cortana.mymoney.ui.base.listener.BaseCallBack;
 import com.vn.hcmute.team.cortana.mymoney.usecase.base.Action;
+import com.vn.hcmute.team.cortana.mymoney.usecase.base.TypeRepository;
 import com.vn.hcmute.team.cortana.mymoney.usecase.base.UseCase;
 import com.vn.hcmute.team.cortana.mymoney.usecase.remote.PersonUseCase.PersonRequest;
 import com.vn.hcmute.team.cortana.mymoney.utils.SecurityUtil;
@@ -53,19 +54,19 @@ public class PersonUseCase extends UseCase<PersonRequest> {
         String action = requestValues.getAction();
         switch (action) {
             case Action.ACTION_GET_PERSON:
-                doGetPerson(requestValues.getCallBack());
+                doGetPerson(requestValues.getCallBack(),requestValues.getTypeRepository());
                 break;
             case Action.ACTION_ADD_PERSON:
                 doAddPerson(requestValues.getCallBack(),
-                          ((CRUDPersonRequest) requestValues).getPerson());
+                          ((CRUDPersonRequest) requestValues).getPerson(),requestValues.getTypeRepository());
                 break;
             case Action.ACTION_REMOVE_PERSON:
                 doRemovePerson(requestValues.getCallBack(),
-                          ((CRUDPersonRequest) requestValues).getPerson().getPersonid());
+                          ((CRUDPersonRequest) requestValues).getPerson().getPersonid(),requestValues.getTypeRepository());
                 break;
             case Action.ACTION_UPDATE_PERSON:
                 doUpdatePerson(requestValues.getCallBack(),
-                          ((CRUDPersonRequest) requestValues).getPerson());
+                          ((CRUDPersonRequest) requestValues).getPerson(),requestValues.getTypeRepository());
                 break;
             case Action.ACTION_SYNC_PERSON:
                 doSyncPerson(requestValues.getCallBack(),
@@ -98,7 +99,6 @@ public class PersonUseCase extends UseCase<PersonRequest> {
         };
         
         if (!this.mCompositeDisposable.isDisposed()) {
-            
             mDisposable = mDataRepository.syncPerson(persons, userid, token)
                       .subscribeOn(Schedulers.io())
                       .observeOn(AndroidSchedulers.mainThread())
@@ -114,15 +114,8 @@ public class PersonUseCase extends UseCase<PersonRequest> {
         }
     }
     
-    private void doUpdatePerson(final BaseCallBack<Object> callBack, Person person) {
-        String userid = mDataRepository.getUserId();
-        String token = mDataRepository.getUserToken();
-        
-        if (TextUtils.isEmpty(userid) || TextUtils.isEmpty(token)) {
-            callBack.onFailure(new UserLoginException(
-                      mContext.getString(R.string.message_warning_need_login)));
-            return;
-        }
+    private void doUpdatePerson(final BaseCallBack<Object> callBack, Person person,TypeRepository typeRepository) {
+       
         
         this.mDisposableSingleObserver = new DisposableSingleObserver<Object>() {
             @Override
@@ -138,18 +131,40 @@ public class PersonUseCase extends UseCase<PersonRequest> {
         };
         
         if (!this.mCompositeDisposable.isDisposed()) {
-            
-            mDisposable = mDataRepository.updatePerson(person, userid, token)
-                      .subscribeOn(Schedulers.io())
-                      .observeOn(AndroidSchedulers.mainThread())
-                      .doOnSubscribe(new Consumer<Disposable>() {
-                          @Override
-                          public void accept(Disposable disposable) throws Exception {
-                              callBack.onLoading();
-                          }
-                      })
-                      .singleOrError()
-                      .subscribeWith(this.mDisposableSingleObserver);
+            if(typeRepository==TypeRepository.REMOTE){
+                String userid = mDataRepository.getUserId();
+                String token = mDataRepository.getUserToken();
+    
+                if (TextUtils.isEmpty(userid) || TextUtils.isEmpty(token)) {
+                    callBack.onFailure(new UserLoginException(
+                              mContext.getString(R.string.message_warning_need_login)));
+                    return;
+                }
+                mDisposable = mDataRepository.updatePerson(person, userid, token)
+                          .subscribeOn(Schedulers.io())
+                          .observeOn(AndroidSchedulers.mainThread())
+                          .doOnSubscribe(new Consumer<Disposable>() {
+                              @Override
+                              public void accept(Disposable disposable) throws Exception {
+                                  callBack.onLoading();
+                              }
+                          })
+                          .singleOrError()
+                          .subscribeWith(this.mDisposableSingleObserver);
+            }else {
+                mDisposable = mDataRepository.updateLocalPerson(person)
+                          .subscribeOn(Schedulers.computation())
+                          .observeOn(AndroidSchedulers.mainThread())
+                          .doOnSubscribe(new Consumer<Disposable>() {
+                              @Override
+                              public void accept(Disposable disposable) throws Exception {
+                                  callBack.onLoading();
+                              }
+                          })
+                          .singleOrError()
+                          .subscribeWith(this.mDisposableSingleObserver);
+            }
+          
             this.mCompositeDisposable.add(mDisposable);
         }
     }
@@ -162,15 +177,8 @@ public class PersonUseCase extends UseCase<PersonRequest> {
         }
     }
     
-    private void doGetPerson(final BaseCallBack<Object> callBack) {
-        String userid = mDataRepository.getUserId();
-        String token = mDataRepository.getUserToken();
-        
-        if (TextUtils.isEmpty(userid) || TextUtils.isEmpty(token)) {
-            callBack.onFailure(new UserLoginException(
-                      mContext.getString(R.string.message_warning_need_login)));
-            return;
-        }
+    private void doGetPerson(final BaseCallBack<Object> callBack,TypeRepository typeRepository) {
+       
         
         this.mDisposableSingleObserver = new DisposableSingleObserver<Object>() {
             @Override
@@ -186,32 +194,47 @@ public class PersonUseCase extends UseCase<PersonRequest> {
         };
         
         if (!this.mCompositeDisposable.isDisposed()) {
-            
-            mDisposable = mDataRepository.getPerson(userid, token)
-                      .subscribeOn(Schedulers.io())
-                      .observeOn(AndroidSchedulers.mainThread())
-                      .doOnSubscribe(new Consumer<Disposable>() {
-                          @Override
-                          public void accept(Disposable disposable) throws Exception {
-                              callBack.onLoading();
-                          }
-                      })
-                      .singleOrError()
-                      .subscribeWith(this.mDisposableSingleObserver);
+            if(typeRepository==TypeRepository.REMOTE){
+                String userid = mDataRepository.getUserId();
+                String token = mDataRepository.getUserToken();
+    
+                if (TextUtils.isEmpty(userid) || TextUtils.isEmpty(token)) {
+                    callBack.onFailure(new UserLoginException(
+                              mContext.getString(R.string.message_warning_need_login)));
+                    return;
+                }
+                mDisposable = mDataRepository.getPerson(userid, token)
+                          .subscribeOn(Schedulers.io())
+                          .observeOn(AndroidSchedulers.mainThread())
+                          .doOnSubscribe(new Consumer<Disposable>() {
+                              @Override
+                              public void accept(Disposable disposable) throws Exception {
+                                  callBack.onLoading();
+                              }
+                          })
+                          .singleOrError()
+                          .subscribeWith(this.mDisposableSingleObserver);
+            }else {
+                String userid = mDataRepository.getUserId();
+                mDisposable = mDataRepository.getLocalListPerson(userid)
+                          .subscribeOn(Schedulers.computation())
+                          .observeOn(AndroidSchedulers.mainThread())
+                          .doOnSubscribe(new Consumer<Disposable>() {
+                              @Override
+                              public void accept(Disposable disposable) throws Exception {
+                                  callBack.onLoading();
+                              }
+                          })
+                          .singleOrError()
+                          .subscribeWith(this.mDisposableSingleObserver);
+            }
+         
             this.mCompositeDisposable.add(mDisposable);
         }
     }
     
-    private void doAddPerson(final BaseCallBack<Object> callBack, Person person) {
-        String userid = mDataRepository.getUserId();
-        String token = mDataRepository.getUserToken();
-        
-        if (TextUtils.isEmpty(userid) || TextUtils.isEmpty(token)) {
-            callBack.onFailure(new UserLoginException(
-                      mContext.getString(R.string.message_warning_need_login)));
-            return;
-        }
-        
+    private void doAddPerson(final BaseCallBack<Object> callBack, Person person,TypeRepository typeRepository) {
+    
         if (TextUtils.isEmpty(person.getName())) {
             callBack.onFailure(
                       new PersonException(mContext.getString(R.string.message_name_person_empty)));
@@ -232,35 +255,53 @@ public class PersonUseCase extends UseCase<PersonRequest> {
         };
         
         if (!this.mCompositeDisposable.isDisposed()) {
-            
-            person.setPersonid(SecurityUtil.getRandomUUID());
-            person.setUserid(userid);
-            
-            mDisposable = mDataRepository.addPerson(person, userid, token)
-                      .subscribeOn(Schedulers.io())
-                      .observeOn(AndroidSchedulers.mainThread())
-                      .doOnSubscribe(new Consumer<Disposable>() {
-                          @Override
-                          public void accept(Disposable disposable) throws Exception {
-                              callBack.onLoading();
-                          }
-                      })
-                      .singleOrError()
-                      .subscribeWith(this.mDisposableSingleObserver);
+            if(typeRepository==TypeRepository.REMOTE){
+                String userid = mDataRepository.getUserId();
+                String token = mDataRepository.getUserToken();
+    
+                if (TextUtils.isEmpty(userid) || TextUtils.isEmpty(token)) {
+                    callBack.onFailure(new UserLoginException(
+                              mContext.getString(R.string.message_warning_need_login)));
+                    return;
+                }
+                
+                person.setPersonid(SecurityUtil.getRandomUUID());
+                person.setUserid(userid);
+    
+                mDisposable = mDataRepository.addPerson(person, userid, token)
+                          .subscribeOn(Schedulers.io())
+                          .observeOn(AndroidSchedulers.mainThread())
+                          .doOnSubscribe(new Consumer<Disposable>() {
+                              @Override
+                              public void accept(Disposable disposable) throws Exception {
+                                  callBack.onLoading();
+                              }
+                          })
+                          .singleOrError()
+                          .subscribeWith(this.mDisposableSingleObserver);
+            }else {
+                String userid = mDataRepository.getUserId();
+                person.setUserid(userid);
+                mDisposable = mDataRepository.addLocalPerson(person)
+                          .subscribeOn(Schedulers.computation())
+                          .observeOn(AndroidSchedulers.mainThread())
+                          .doOnSubscribe(new Consumer<Disposable>() {
+                              @Override
+                              public void accept(Disposable disposable) throws Exception {
+                                  callBack.onLoading();
+                              }
+                          })
+                          .singleOrError()
+                          .subscribeWith(this.mDisposableSingleObserver);
+            }
+         
             this.mCompositeDisposable.add(mDisposable);
             
         }
     }
     
-    private void doRemovePerson(final BaseCallBack<Object> callBack, String personid) {
-        String userid = mDataRepository.getUserId();
-        String token = mDataRepository.getUserToken();
-        
-        if (TextUtils.isEmpty(userid) || TextUtils.isEmpty(token)) {
-            callBack.onFailure(new UserLoginException(
-                      mContext.getString(R.string.message_warning_need_login)));
-            return;
-        }
+    private void doRemovePerson(final BaseCallBack<Object> callBack, String personid,TypeRepository typeRepository) {
+       
         
         this.mDisposableSingleObserver = new DisposableSingleObserver<Object>() {
             @Override
@@ -276,20 +317,41 @@ public class PersonUseCase extends UseCase<PersonRequest> {
         };
         
         if (!this.mCompositeDisposable.isDisposed()) {
-            
-            mDisposable = mDataRepository.removePerson(userid, token, personid)
-                      .subscribeOn(Schedulers.io())
-                      .observeOn(AndroidSchedulers.mainThread())
-                      .doOnSubscribe(new Consumer<Disposable>() {
-                          @Override
-                          public void accept(Disposable disposable) throws Exception {
-                              callBack.onLoading();
-                          }
-                      })
-                      .singleOrError()
-                      .subscribeWith(this.mDisposableSingleObserver);
+            if(typeRepository==TypeRepository.REMOTE){
+                
+                String userid = mDataRepository.getUserId();
+                String token = mDataRepository.getUserToken();
+    
+                if (TextUtils.isEmpty(userid) || TextUtils.isEmpty(token)) {
+                    callBack.onFailure(new UserLoginException(
+                              mContext.getString(R.string.message_warning_need_login)));
+                    return;
+                }
+                mDisposable = mDataRepository.removePerson(userid, token, personid)
+                          .subscribeOn(Schedulers.io())
+                          .observeOn(AndroidSchedulers.mainThread())
+                          .doOnSubscribe(new Consumer<Disposable>() {
+                              @Override
+                              public void accept(Disposable disposable) throws Exception {
+                                  callBack.onLoading();
+                              }
+                          })
+                          .singleOrError()
+                          .subscribeWith(this.mDisposableSingleObserver);
+            }else {
+                mDisposable = mDataRepository.deleteLocalPeron(personid)
+                          .subscribeOn(Schedulers.computation())
+                          .observeOn(AndroidSchedulers.mainThread())
+                          .doOnSubscribe(new Consumer<Disposable>() {
+                              @Override
+                              public void accept(Disposable disposable) throws Exception {
+                                  callBack.onLoading();
+                              }
+                          })
+                          .singleOrError()
+                          .subscribeWith(this.mDisposableSingleObserver);
+            }
             this.mCompositeDisposable.add(mDisposable);
-            
         }
     }
     
@@ -298,6 +360,7 @@ public class PersonUseCase extends UseCase<PersonRequest> {
         
         private String action;
         private BaseCallBack<Object> callBack;
+        private TypeRepository typeRepository;
         
         public PersonRequest(@NonNull String action, @Nullable BaseCallBack<Object> callBack) {
             this.action = action;
@@ -318,6 +381,13 @@ public class PersonUseCase extends UseCase<PersonRequest> {
         
         public void setCallBack(BaseCallBack<Object> callBack) {
             this.callBack = callBack;
+        }
+        public TypeRepository getTypeRepository() {
+            return typeRepository;
+        }
+        public void setTypeRepository(
+                  TypeRepository typeRepository) {
+            this.typeRepository = typeRepository;
         }
     }
